requirements.txt
annotated-types==0.7.0
anyio==4.4.0
backoff==2.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
filelock==3.15.4
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.41
h11==0.14.0
haystack-ai==2.3.1
haystack-experimental==0.1.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
lazy-imports==0.3.1
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
openai==1.38.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
posthog==3.5.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
setuptools==69.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.43.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wheel==0.43.0
